backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.24.0
pydantic==2.8.2
pydantic-settings==2.3.4
python-multipart==0.0.9
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
beanie==1.23.6
python-docx==1.1.2
pypdf==4.2.0
openai==1.35.13
# passlib[bcrypt]==1.7.4  # Removed - using direct bcrypt instead
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
